+++ src/components/Modals/ConfirmationDeleteNode/ConfirmationDeleteNode.js
