+++ src/components/Modals/ConfirmationDeleteNode/index.js
