+++ src/components/Modals/ConfirmationDeleteNode/useDeleteNodeModal.js
