+++ src/components/Modals/ConfirmationDeleteStructure/ConfirmationDeleteStructure.js
