+++ src/components/Modals/ConfirmationDeleteStructure/ConfirmationDeleteStructure.module.css
