+++ src/components/Modals/ConfirmationDeleteStructure/index.js
