+++ src/components/Modals/ConfirmationDeleteStructure/useDeleteStructureModal.js
